x86/ioapic: sanitize the IO-APIC pins before enabling the local APIC
The current logic to init the local APIC and the IO-APIC does init the
former first before doing any kind of sanitation on the IO-APIC pin
confirmation.  It's already noted on enable_IO_APIC() that Xen
shouldn't trust the IO-APIC being empty at bootup.

At XenServer we have a system where the IO-APIC is handed to Xen with
IO-APIC 0 pin 0 unmasked, set to Fixed delivery mode and with a vector
of 0.  Once the local APIC is enabled periodic interrupts from such
pin (likely connected to the PIT) cause a storm of errors:

APIC error on CPU0: 40(40), Received illegal vector

That prevent Xen from booting.

Fix this by moving the masking of IO-APIC pins ahead of the enabling
of the local APIC.  Note that before doing such masking Xen attempts
to detect the pin where the legacy i8259 is connected, and that logic
relies on the pin being unmasked, hence the logic is also moved ahead
of enabling the local APIC.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>