x86/spec-ctrl: don't enable IBRS if IBPB-on-entry and STIBP are enabled
IBRS has slightly different properties, depending on whether the CPU vendor is
Intel or AMD.

For Intel IBRS ensures that predicted targets of indirect
branches executed in that predictor mode with IBRS = 1 cannot be controlled by
software that was executed in a less privileged predictor mode.

For AMD IBRS is even more restrictive, and prevents speculation at the
predicted target of any instruction that is decoded as an indirect branch,
regardless of the predicted branch type.

STIBP restricts the sharing of branch prediction between logical processors on
a core.

Finally, IBPB-on-entry prevents software that executed before the barrier from
controlling the predicted targets of indirect branches executed after the
barrier on the same logical processor.

When enabling for IBPB-on-entry and STIBP it's impossible for a guest to
control the predicted targets of indirect branches, since IBPB-on-entry
prevents software executed before the barrier from doing so, and STIBP prevents
a different thread on the same core from controlling the predicted targets.

Avoid the overhead of IBRS when IBPB-on-entry and STIBP are both enabled.

Suggested-by: Andrew Cooper <andrew.cooper3@citrix.com>
Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>