x86/spec-ctrl: avoid modifying opt_ibrs when CET is in use
Do not overwrite the value of opt_ibrs, even when set as '-1' (default).
Not overwriting allows further code to distinguish whether the user has forced
any specific selection of IBRS.  This requires shuffling a bit the code in
order to make sure the local ibrs and thunk variables are always set.

No functional change intended.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>