alternatives: remove the executable section flag and put in .init.data
The alternative binary text blobs replacements (the contents of the
.altinstr_replacement section) don't need to have the execute
section flag set, neither need to live in .text related sections, since the
alternative replacement blocks are never executed from that location.  Instead
such blocks are copied over localtions in .text sections, that already have the
appropriate placement and section flags.

Adjust the flags of the .altinstr_replacement to drop the executable flag, and
also change the position in the linker script so that the replacements are
packed in .init.data instead of .init.text (even better would be .init.rodata
if we had it).

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>