alternatives: remove executable section flag and put replacements in .init.data
The alternative binary text blobs replacements (the contents of the
.altinstr_replacement section) don't need the executable section flag set,
neither need to be placed in .text section, because the alternative replacement
blocks are never executed from that location.  Instead such blocks are copied
over locations in .text sections that are executable.

Adjust the flags of the .altinstr_replacement to drop the executable flag, and
also change the position in the linker script so that the replacements are
packed in .init.data instead of .init.text (even better would be .init.rodata
if we had it).

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>